moving the loadFarmers func into a MapViewService
